backend/requirements.txt
fastapi
uvicorn
opencv-python
torch
torchvision
ultralytics
easyocr
moviepy
ffmpeg-python
python-multipart
yt-dlp